requirements.txt
pillow>=9.0.0
pillow-heif>=0.10.0 